requirements.txt
pandas==1.5.0
Pillow==9.3.0
streamlit==1.13.0
streamlit_option_menu==0.3.2
